Added ant_setup
Started playing around with the ant_setup file.  Have managed to get ants to obey boundries of the wall, and a kind of phermone trail being left.  But have an issue with only being able to spawn ants outside of the walls.  Need to figure out how to spawn them inside.